复习结束力！

+++ mc-/mc-/role_operate.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <vector>
-void steve_make(std::vector<int>&cde);
-void move(std::vector<std::vector<std::vector<int>>>&map_id, std::vector<box>& box_data, int& x, int& y, int& z);
+void steve_make(std::vector<double>&cde);
+void move(std::vector<std::vector<std::vector<int>>>&map_id, std::vector<box>& box_data, double& x, double& y, double& z);
